embedding_db/requirements.txt
openai>=1.40.0
faiss-cpu>=1.8.0
numpy>=1.26.4
pandas>=2.2.2
python-dotenv>=1.0.1
tqdm>=4.66.4
sqlite-utils>=3.37
